Add Overcooked items management functions
- Introduced overcooked_items.h to manage item and cooking pot functionalities within the Overcooked environment.
- Implemented functions for retrieving, adding, and removing items, as well as initializing cooking pots and updating their cooking states.
- Updated overcooked.h to include the new overcooked_items.h header, removing redundant function declarations.
